rename: HeaderLecturerControls를 HeaderInstructorControls로 변경
HeaderLecturerControls컴포넌트의 이름을 HeaderInstructorControls로 변경했습니다.
